+++ src/app/basicauction/page.tsx
@@ -6,6 +6,7 @@ import AuctionDropDown from '@/app/basicauction/components/auctiondropdown';
 import Checkbox from '@/app/basicauction/components/checkbox';
 import LeftSection from '@/components/LeftSection';
 import MaxLayout from '@/components/MaxLayout';
+import { flattenCategories } from '@/utils/flattenCategoriesTree';
 import getMetadata from '@/utils/getMetadata';
 
 import * as S from './Basicauction.css';
@@ -18,14 +19,19 @@ interface GenerateMetadataProps {
 export const generateMetadata = async ({
   searchParams,
 }: GenerateMetadataProps): Promise<Metadata> => {
-  const queryValue = searchParams.categoryId;
+  const queryValue = searchParams.categoryId || 'ALL';
+  const categoryListTree = await fetch(
+    `${process.env.NEXT_PUBLIC_SERVER_API_URL}/api/v2/categories/${queryValue}`,
+  )
+    .then(res => res.json())
+    .then(jsonRes => jsonRes.result_data);
 
-  // api 호출해서 카테고리 이름 가져오기
-  // 맨 아래있는 것으로 가져오기
-  // 사진은 배포 하고나 볼 수 있는 듯..
+  const flattenCategoriesList = flattenCategories(categoryListTree);
+  const categoryName = flattenCategoriesList[flattenCategoriesList.length - 1].name;
 
   return getMetadata({
-    title: `${queryValue}`, // TODO 제일 하단의 요소로 가져와야함.
+    title: `${categoryName}`,
+    // TODO 예쁜 사진이 있다면 그것으로 카테고리를 나타내서 openGraphImage를 설정하면 좋을 듯함.
   });
 };
