+++ src/app/layout.tsx
@@ -15,6 +15,7 @@ import usePrefetchQueryWithCookie from '@/hooks/usePrefetchQueryWithCookie';
 import { AuthProvider } from '@/provider/authProvider';
 import TanstackProviders from '@/provider/tanstackProviders';
 import { ToastProvider } from '@/provider/toastProvider';
+import getMetadata from '@/utils/getMetadata';
 
 import type { Metadata, Viewport } from 'next';
 
@@ -24,12 +25,16 @@ const roboto = Roboto({
   display: 'swap',
 });
 
-export const metadata: Metadata = {
-  title: 'Omocha',
-  description: 'Hello, Omocha!',
-  icons: {
-    icon: './icon.ico',
-  },
+// export const metadata: Metadata = {
+//   title: 'Omocha',
+//   description: 'Hello, Omocha!',
+//   icons: {
+//     icon: './icon.ico',
+//   },
+// };
+
+export const generateMetadata = async (): Promise<Metadata> => {
+  return getMetadata();
 };
 
 export const viewport: Viewport = {
